Fix float-equal and double-promotion warnings

+++ DataFormats/interface/KTrack.h
@@ -126,7 +126,7 @@ struct KTrack : public KLV
 				orthog[1] = -p4.x();
 				orthog[2] = 0;
 
-				float vtxErr2 = ROOT::Math::Similarity(pv->covariance, orthog) / p4.Perp2();
+				double vtxErr2 = ROOT::Math::Similarity(pv->covariance, orthog) / p4.Perp2();
 				return getDxy(pv) / sqrt(errDxy*errDxy + vtxErr2);
 				break;
 		}
